// Qanat.Web/src/app/shared/generated/model/parcel-history-dto.ts
export class ParcelHistoryDto { 
    ParcelHistoryID?: number;
    GeographyID?: number;
    ParcelID?: number;
    EffectiveYear?: number;
    UpdateDate?: string;
    UpdateUserID?: number;
    UpdateUserFullName?: string;
    ParcelArea?: number;
    OwnerName?: string;
    OwnerAddress?: string;
    ParcelStatusID?: number;
    IsReviewed?: boolean;
    IsManualOverride?: boolean;
    ReviewDate?: string;
    WaterAccountID?: number;
    WaterAccount?: WaterAccountDisplayDto;
    ParcelStatus?: ParcelStatusSimpleDto;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}